Replace setContent callback, hide on keydown
Previously, I provided a fix for #19. This is a better fix. I realized that the
onBlur callback *does* need to be called on the setContent event.

This ensures the label is hidden if the user it typing (on 'keydown'). This
prevents the label from overlaying with text the user is typing.